probing and tracking full functionality
That concludes the comms console, whee

+++ code/modules/shiptoship/ship/probe_control.dm
@@ -19,13 +19,67 @@
 	var/ship_name = "none" //This has to match the name of the ship its on, as saved on its Mission Control structure.
 	var/probe_loaded = 0
 
+/obj/structure/ship_elements/probe_launcher/proc/LaunchContent()
+	emoteas("Launches its contets with a loud swoosh!")
+	playsound(src, 'sound/effects/bamf.ogg', 25)
+	probe_loaded = 0
+
 /obj/structure/ship_elements/probe_launcher/attackby(obj/item/W, mob/user)
+	if(!(istype(W, /obj/item/ship_probe)))
+		to_chat(usr, SPAN_WARNING("You have no idea how to use these two together."))
+		return
 	if(istype(W, /obj/item/ship_probe))
 		if(probe_loaded == 1)
 			to_chat(usr, SPAN_WARNING("A probe is already loaded."))
+			return
 		if(probe_loaded == 0)
 			usr.visible_message(SPAN_INFO("[usr] starts to reload the probe launcher."), SPAN_INFO("You start to reload the probe launcher"), SPAN_WARNING("You hear loud, mechanic clicking!"))
 			if(do_after(usr, 20, INTERRUPT_ALL|BEHAVIOR_IMMOBILE, BUSY_ICON_BUILD))
 				usr.visible_message(SPAN_INFO("[usr] reloads the probe launcher."), SPAN_INFO("The probe slides into the tube and an audible click is heard."), SPAN_WARNING("The clicking stops after a beep."))
 				probe_loaded = 1
 				qdel(W)
+				return
+	return
+
+/obj/item/ship_tacker
+	name = "PHA-1 type tracker device"
+	desc = "A drone sized device encased in polymer glass with slightly glowing, greenish electronic parts visible inside."
+	desc_lore = "PHA-1 Trackers are EYE-7 probes that have undergone further modification and integration into the PST's systems and offer a permanent window for the Mission Control system to any entity they manage to attach themselves to. This means that while they do not feed any other information, these devices will always feed their location back to the system, allowing for live tracking."
+	icon = 'icons/sectorpatrol/ship/probe.dmi'
+	icon_state = "tracker"
+	var/ship_id = "none"
+	flags_item = NOBLUDGEON
+
+/obj/structure/ship_elements/tracker_launcher
+	name = "PHA-1 tracker launcher system"
+	desc = "A long tube connected to what looks like a simple loading port and a status indicator."
+	desc_lore = "Like the trackers themselves, the PHA-1 tracker launchers are modified EYE-7 drone stations which like the probes takes relative data and turns it into the coordinate-based information that the PST's Mission Control system relies on."
+	icon = 'icons/sectorpatrol/ship/probe96.dmi'
+	icon_state = "tracker_launcher"
+	anchored = 1
+	opacity = 1
+	density = 1
+	var/ship_name = "none" //This has to match the name of the ship its on, as saved on its Mission Control structure.
+	var/tracker_loaded = 0
+
+/obj/structure/ship_elements/tracker_launcher/proc/LaunchContent()
+	emoteas("Launches its contets with a loud swoosh!")
+	playsound(src, 'sound/effects/bamf.ogg', 25)
+	tracker_loaded = 0
+
+/obj/structure/ship_elements/tracker_launcher/attackby(obj/item/W, mob/user)
+	if(!(istype(W, /obj/item/ship_tacker)))
+		to_chat(usr, SPAN_WARNING("You have no idea how to use these two together."))
+		return
+	if(istype(W, /obj/item/ship_tacker))
+		if(tracker_loaded == 1)
+			to_chat(usr, SPAN_WARNING("A tracker is already loaded."))
+			return
+		if(tracker_loaded == 0)
+			usr.visible_message(SPAN_INFO("[usr] starts to reload the tracker launcher."), SPAN_INFO("You start to reload the tracker launcher"), SPAN_WARNING("You hear loud, mechanic clicking!"))
+			if(do_after(usr, 20, INTERRUPT_ALL|BEHAVIOR_IMMOBILE, BUSY_ICON_BUILD))
+				usr.visible_message(SPAN_INFO("[usr] reloads the tracker launcher."), SPAN_INFO("The tracker slides into the tube and an audible click is heard."), SPAN_WARNING("The clicking stops after a beep."))
+				tracker_loaded = 1
+				qdel(W)
+				return
+	return
